+++ satellite_propagation/matrix_operations.c
@@ -3,7 +3,7 @@
 
 
 
-void mult_matrix(double a[3][3], double b[3][3], double c[3][3])
+void mult_matrix(long double a[3][3], long double b[3][3], long double c[3][3])
 {
     int i, j, k;
     for (i=0; i<3; i++)
@@ -20,7 +20,7 @@ void mult_matrix(double a[3][3], double b[3][3], double c[3][3])
 }
 
 
-void mult_matrix_by_vector(double matr[3][3], double vec[3], double res[3])
+void mult_matrix_by_vector(long double matr[3][3], long double vec[3], long double res[3])
 {
     int i, j;
     for (i=0; i<3; i++)
@@ -37,20 +37,20 @@ void mult_matrix_by_vector(double matr[3][3], double vec[3], double res[3])
 void transpose(void *dest, void *src, int src_h, int src_w)
 {
     int i, j;
-    double (*d)[src_h] = dest, (*s)[src_w] = src;
+    long double (*d)[src_h] = dest, (*s)[src_w] = src;
     for (i = 0; i < src_h; i++)
         for (j = 0; j < src_w; j++)
             d[j][i] = s[i][j];
 }
 
 
-void print_matrix(int n, int m, double *matrix)
+void print_matrix(int n, int m, long double *matrix)
 {
     for (int i=0; i<n; i++)
     {
         for (int j=0;j<m;j++)
         {
-            printf("%2.15f   ", *((matrix+i*n) + j));
+            printf("%2.15Lf   ", *((matrix+i*n) + j));
         }
         printf("\n");
     }
